Move CUDA to a package extension
