+++ src/Architectures.jl
@@ -4,7 +4,6 @@ export AbstractArchitecture
 export CPU, GPU, MultiGPU
 export device, architecture, array_type, arch_array, unified_array, device_copy_to!
 
-using CUDA
 using KernelAbstractions
 using Adapt
 using OffsetArrays
@@ -32,18 +31,18 @@ Run Oceananigans on a single NVIDIA CUDA GPU.
 struct GPU <: AbstractArchitecture end
 
 #####
-##### These methods are extended in Distributed.jl
+##### These methods are extended in Distributed.jl and CUDAExt.jl
 #####
 
 device(::CPU) = KernelAbstractions.CPU()
-device(::GPU) = CUDA.CUDABackend(; always_inline=true)
 
 architecture() = nothing
 architecture(::Number) = nothing
 architecture(::Array) = CPU()
-architecture(::CuArray) = GPU()
 architecture(a::SubArray) = architecture(parent(a))
 architecture(a::OffsetArray) = architecture(parent(a))
+
+available(::CPU) = true
 
 """
     child_architecture(arch)
@@ -54,19 +53,9 @@ On single-process, non-distributed systems, return `arch`.
 child_architecture(arch) = arch
 
 array_type(::CPU) = Array
-array_type(::GPU) = CuArray
 
-arch_array(::CPU, a::Array)   = a
-arch_array(::CPU, a::CuArray) = Array(a)
-arch_array(::GPU, a::Array)   = CuArray(a)
-arch_array(::GPU, a::CuArray) = a
-
-arch_array(::GPU, a::SubArray{<:Any, <:Any, <:CuArray}) = a
-arch_array(::CPU, a::SubArray{<:Any, <:Any, <:CuArray}) = Array(a)
-
-arch_array(::GPU, a::SubArray{<:Any, <:Any, <:Array}) = CuArray(a)
+arch_array(::CPU, a::Array) = a
 arch_array(::CPU, a::SubArray{<:Any, <:Any, <:Array}) = a
-
 arch_array(arch, a::AbstractRange) = a
 arch_array(arch, a::OffsetArray) = OffsetArray(arch_array(arch, a.parent), a.offsets...)
 arch_array(arch, ::Nothing)   = nothing
@@ -74,32 +63,9 @@ arch_array(arch, a::Number)   = a
 arch_array(arch, a::Function) = a
 
 unified_array(::CPU, a) = a
-unified_array(::GPU, a) = a
-
-function unified_array(::GPU, arr::AbstractArray) 
-    buf = Mem.alloc(Mem.Unified, sizeof(arr))
-    vec = unsafe_wrap(CuArray{eltype(arr),length(size(arr))}, convert(CuPtr{eltype(arr)}, buf), size(arr))
-    finalizer(vec) do _
-        Mem.free(buf)
-    end
-    copyto!(vec, arr)
-    return vec
-end
-
-## Only for contiguous data!! (i.e. only if the offset for pointer(dst::CuArrat, offset::Int) is 1)
-@inline function device_copy_to!(dst::CuArray, src::CuArray; async::Bool = false) 
-    n = length(src)
-    context!(context(src)) do
-        GC.@preserve src dst begin
-            unsafe_copyto!(pointer(dst, 1), pointer(src, 1), n; async)
-        end
-    end
-    return dst
-end
  
 @inline device_copy_to!(dst::Array, src::Array; kw...) = Base.copyto!(dst, src)
 
-@inline unsafe_free!(a::CuArray) = CUDA.unsafe_free!(a)
 @inline unsafe_free!(a)          = nothing
 
 end # module
