+++ src/Fields/field.jl
@@ -2,6 +2,7 @@ using Oceananigans.BoundaryConditions: OBC, MCBC
 using Oceananigans.Grids: parent_index_range, index_range_offset, default_indices, all_indices, validate_indices
 
 using Adapt
+import GPUArraysCore
 using KernelAbstractions: @kernel, @index
 using Base: @propagate_inbounds
 
@@ -648,7 +649,7 @@ for reduction in (:sum, :maximum, :minimum, :all, :any, :prod)
             Base.$(reduction!)(identity, r, conditioned_c, init=false)
 
             if dims isa Colon
-                return CUDA.@allowscalar first(r)
+                return GPUArraysCore.@allowscalar first(r)
             else
                 return r
             end
@@ -687,7 +688,7 @@ Statistics.mean!(r::ReducedField, a::AbstractArray; kwargs...) = Statistics.mean
 function Statistics.norm(a::AbstractField; condition = nothing)
     r = zeros(a.grid, 1)
     Base.mapreducedim!(x -> x * x, +, r, condition_operand(a, condition, 0))
-    return CUDA.@allowscalar sqrt(r[1])
+    return GPUArraysCore.@allowscalar sqrt(r[1])
 end
 
 function Base.isapprox(a::AbstractField, b::AbstractField; kw...)
