+++ src/Grids/Grids.jl
@@ -17,8 +17,7 @@ export minimum_xspacing, minimum_yspacing, minimum_zspacing
 export offset_data, new_data
 export on_architecture
 
-using CUDA
-using CUDA: has_cuda
+using GPUArraysCore
 using Adapt
 using OffsetArrays
