+++ src/Grids/grid_generation.jl
@@ -6,11 +6,11 @@
 get_domain_extent(::Nothing, N)             = (1, 1)
 get_domain_extent(coord, N)                 = (coord[1], coord[2])
 get_domain_extent(coord::Function, N)       = (coord(1), coord(N+1))
-get_domain_extent(coord::AbstractVector, N) = CUDA.@allowscalar (coord[1], coord[N+1])
+get_domain_extent(coord::AbstractVector, N) = GPUArraysCore.@allowscalar (coord[1], coord[N+1])
 
 get_coord_face(coord::Nothing, i) = 1
 get_coord_face(coord::Function, i) = coord(i)
-get_coord_face(coord::AbstractVector, i) = CUDA.@allowscalar coord[i]
+get_coord_face(coord::AbstractVector, i) = GPUArraysCore.@allowscalar coord[i]
 
 const AT = AbstractTopology
 lower_exterior_Δcoordᶠ(::AT, Fi, Hcoord) = [Fi[end - Hcoord + i] - Fi[end - Hcoord + i - 1] for i = 1:Hcoord]
