+++ src/Solvers/fft_based_poisson_solver.jl
@@ -108,7 +108,7 @@ function solve!(ϕ, solver::FFTBasedPoissonSolver, b, m=0)
     # If m === 0, the "zeroth mode" at `i, j, k = 1, 1, 1` is undetermined;
     # we set this to zero by default. Another slant on this "problem" is that
     # λx[1, 1, 1] + λy[1, 1, 1] + λz[1, 1, 1] = 0, which yields ϕ[1, 1, 1] = Inf or NaN.
-    m === 0 && CUDA.@allowscalar ϕc[1, 1, 1] = 0
+    m === 0 && GPUArraysCore.@allowscalar ϕc[1, 1, 1] = 0
 
     # Apply backward transforms in order
     [transform!(ϕc, solver.buffer) for transform! in solver.transforms.backward]
